+++ src/MIToS.jl
@@ -6,6 +6,7 @@ module MIToS
 
   include(joinpath("Utils", "Utils.jl"))
   include(joinpath("MSA", "MSA.jl"))
+  include(joinpath("Pfam", "Pfam.jl"))
   include(joinpath("Clustering", "Clustering.jl"))
   include(joinpath("Information", "Information.jl"))
   include(joinpath("PDB", "PDB.jl"))