+++ src/Pfam/Pfam.jl
@@ -1,0 +1,11 @@
+module Pfam
+
+  using MIToS.Utils
+  using MIToS.MSA
+
+  export Stockholm, downloadpfam, getseq2pdb
+
+  include("download.jl")
+  include("pdb.jl")
+
+end